requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==6.1.0
build==1.0.3
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorlog==4.8.0
comm==0.2.0
contourpy==1.2.0
crashtest==0.4.1
cryptography==41.0.7
cssutils==2.9.0
cycler==0.12.1
Cython==3.0.6
dacite==1.8.1
dataframe-image==0.2.3
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
dulwich==0.21.7
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.104.1
fastjsonschema==2.19.0
filelock==3.13.1
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.12.1
gcsfs==2023.12.1
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.15.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-cloud-core==2.4.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
h11==0.14.0
html2image==2.0.4.3
hydra-core==1.3.2
idna==3.6
importlib-metadata==7.0.0
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.27.1
ipython==8.18.1
jaraco.classes==3.3.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab_pygments==0.3.0
keyring==24.3.0
kiwisolver==1.4.5
lxml==4.9.3
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.4
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
numpy==1.26.2
oauthlib==3.2.2
omegaconf==2.3.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.3.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prdc==0.2
prompt-toolkit==3.0.41
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pyparsing==3.1.1
pyproject_hooks==1.0.0
pysen==0.10.5
PySocks==1.7.1
pytest==7.4.3
python-dateutil==2.8.2
pytorch-fid==0.3.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.5.2
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rpds-py==0.13.2
rsa==4.9
SciencePlots==2.1.1
scikit-learn==1.0.2
scipy==1.11.4
seaborn==0.13.0
SecretStorage==3.3.3
selenium==4.16.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.27.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.3
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1+cu117
torch-tb-profiler==0.4.3
torch_geometric==2.4.0
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.37.2
trio==0.23.1
trio-websocket==0.11.1
trove-classifiers==2023.11.29
typing_extensions==4.8.0
tzdata==2023.3
unidiff==0.7.5
urllib3==2.1.0
uvicorn==0.24.0.post1
virtualenv==20.25.0
wcwidth==0.2.12
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
